updated port

+++ testing-tools/src/types/config/node.rs
@@ -72,7 +72,7 @@ impl<C: GlobalTestingConfig> TestingNodeConfig<C> {
 
         Anvil::new()
             .chain_id(1)
-            .port(8545 + self.node_id)
+            .port((8545 + self.node_id) as u16)
             .fork(self.global_config.eth_ws_url())
             .arg("--ipc")
             .arg(self.global_config.anvil_rpc_endpoint(self.node_id))
@@ -85,7 +85,7 @@ impl<C: GlobalTestingConfig> TestingNodeConfig<C> {
     fn configure_devnet_anvil(&self) -> Anvil {
         let mut anvil_builder = Anvil::new()
             .chain_id(1)
-            .port(8545 + self.node_id)
+            .port((8545 + self.node_id) as u16)
             .arg("--ipc")
             .arg(self.global_config.anvil_rpc_endpoint(self.node_id))
             .arg("--code-size-limit")
